fix teleport gets rendered to nonexistent target
