+++ packages/web-runtime/src/components/AppFloatingActionButton.vue
@@ -1,48 +1,22 @@
 <template>
-  <template v-if="floatingActionButton">
-    <div v-if="!isMobile">
-      <teleport to="#web-nav-sidebar-floating-action-button">
-        <div class="pb-3 px-2">
-          <oc-button
-            :id="getButtonId(floatingActionButton.id)"
-            :disabled="isDisabled"
-            appearance="filled"
-            class="oc-app-floating-action-button w-full"
-            @click="floatingActionButton.handler?.()"
-          >
-            <oc-icon :name="floatingActionButton.icon" />
-            <span v-text="floatingActionButton.label()" />
-          </oc-button>
-          <template
-            v-if="floatingActionButton.dropComponent && floatingActionButton.mode() === 'drop'"
-          >
-            <component
-              :is="floatingActionButton.dropComponent"
-              :toggle="`#${getButtonId(floatingActionButton.id)}`"
-            />
-          </template>
-        </div>
-      </teleport>
-    </div>
-    <template v-else-if="!isDisabled">
-      <oc-floating-action-button
-        :button-id="getButtonId(floatingActionButton.id)"
-        class="oc-app-floating-action-button"
-        mode="action"
-        :handler="floatingActionButton.handler"
+  <template v-if="floatingActionButton && isMobile && !floatingActionButton?.isDisabled?.()">
+    <oc-floating-action-button
+      :button-id="getButtonId(floatingActionButton.id)"
+      class="oc-app-floating-action-button"
+      mode="action"
+      :handler="floatingActionButton.handler"
+    />
+    <template v-if="floatingActionButton.dropComponent && floatingActionButton.mode() === 'drop'">
+      <component
+        :is="floatingActionButton.dropComponent"
+        :toggle="`#${getButtonId(floatingActionButton.id)}`"
       />
-      <template v-if="floatingActionButton.dropComponent && floatingActionButton.mode() === 'drop'">
-        <component
-          :is="floatingActionButton.dropComponent"
-          :toggle="`#${getButtonId(floatingActionButton.id)}`"
-        />
-      </template>
     </template>
   </template>
 </template>
 
 <script setup lang="ts">
-import { computed, onBeforeUnmount, ref, unref, watchEffect } from 'vue'
+import { computed, ref, unref } from 'vue'
 import {
   FloatingActionButtonExtension,
   useActiveApp,
@@ -63,28 +37,7 @@ const floatingActionButton = computed(() => {
   }).find(({ isVisible }) => !isVisible || isVisible())
 })
 
-const getButtonId = (extensionId: string): string => {
-  const prefix = unref(isMobile) ? 'mobile-' : ''
-  return `${prefix}app-floating-action-button-${extensionId.replace(/\./g, '-')}`
+function getButtonId(extensionId: string): string {
+  return `mobile-app-floating-action-button-${extensionId.replace(/\./g, '-')}`
 }
-
-let timeoutId: ReturnType<typeof setTimeout> | null = null
-
-// use a timeout to avoid flickering of the button in case the isDisabled state changes rapidly
-watchEffect(() => {
-  const disabled = unref(floatingActionButton)?.isDisabled?.() ?? false
-  if (timeoutId) {
-    clearTimeout(timeoutId)
-  }
-  timeoutId = setTimeout(() => {
-    isDisabled.value = disabled
-    timeoutId = null
-  }, 50)
-})
-
-onBeforeUnmount(() => {
-  if (timeoutId) {
-    clearTimeout(timeoutId)
-  }
-})
 </script>
